+++ includes/blocks/class-container.php
@@ -160,6 +160,7 @@ class GenerateBlocks_Block_Container {
 			'fontSizeMobile' => '',
 			'fontSizeUnit' => 'px',
 			'textTransform' => '',
+			'useLegacyLayout' => false,
 		];
 	}
 
@@ -235,7 +236,11 @@ class GenerateBlocks_Block_Container {
 		$css->add_property( 'text-transform', $settings['textTransform'] );
 		$css->add_property( 'margin', array( $settings['marginTop'], $settings['marginRight'], $settings['marginBottom'], $settings['marginLeft'] ), $settings['marginUnit'] );
 
-		if ( 'contained' === $settings['outerContainer'] && ! $settings['isGrid'] ) {
+		if ( ! $settings['useLegacyLayout'] ) {
+			$css->add_property( 'padding', array( $settings['paddingTop'], $settings['paddingRight'], $settings['paddingBottom'], $settings['paddingLeft'] ), $settings['paddingUnit'] );
+		}
+
+		if ( $settings['useLegacyLayout'] && 'contained' === $settings['outerContainer'] && ! $settings['isGrid'] ) {
 			if ( ! empty( $containerWidth ) ) {
 				$css->add_property( 'max-width', absint( $containerWidth ), 'px' );
 				$css->add_property( 'margin-left', 'auto' );
@@ -342,26 +347,28 @@ class GenerateBlocks_Block_Container {
 			}
 		}
 
-		$css->set_selector( '.gb-container-' . $id . ' > .gb-inside-container' );
-		$css->add_property( 'padding', array( $settings['paddingTop'], $settings['paddingRight'], $settings['paddingBottom'], $settings['paddingLeft'] ), $settings['paddingUnit'] );
+		if ( $settings['useLegacyLayout'] ) {
+			$css->set_selector( '.gb-container-' . $id . ' > .gb-inside-container' );
+			$css->add_property( 'padding', array( $settings['paddingTop'], $settings['paddingRight'], $settings['paddingBottom'], $settings['paddingLeft'] ), $settings['paddingUnit'] );
 
-		if ( 'contained' === $settings['innerContainer'] && ! $settings['isGrid'] ) {
-			if ( ! empty( $containerWidth ) ) {
-				$css->add_property( 'max-width', absint( $containerWidth ), 'px' );
-				$css->add_property( 'margin-left', 'auto' );
-				$css->add_property( 'margin-right', 'auto' );
+			if ( 'contained' === $settings['innerContainer'] && ! $settings['isGrid'] ) {
+				if ( ! empty( $containerWidth ) ) {
+					$css->add_property( 'max-width', absint( $containerWidth ), 'px' );
+					$css->add_property( 'margin-left', 'auto' );
+					$css->add_property( 'margin-right', 'auto' );
+				}
 			}
-		}
 
-		if ( $usingMinHeightFlex ) {
-			$css->add_property( 'width', '100%' );
+			if ( $usingMinHeightFlex ) {
+				$css->add_property( 'width', '100%' );
 
-			$usingMinHeightInnerWidth = true;
-		}
+				$usingMinHeightInnerWidth = true;
+			}
 
-		if ( $innerZIndex || 0 === $innerZIndex ) {
-			$css->add_property( 'z-index', $innerZIndex );
-			$css->add_property( 'position', 'relative' );
+			if ( $innerZIndex || 0 === $innerZIndex ) {
+				$css->add_property( 'z-index', $innerZIndex );
+				$css->add_property( 'position', 'relative' );
+			}
 		}
 
 		$css->set_selector( '.gb-container-' . $id . ' a, .gb-container-' . $id . ' a:visited' );
@@ -463,6 +470,10 @@ class GenerateBlocks_Block_Container {
 		$tablet_css->add_property( 'border-width', array( $settings['borderSizeTopTablet'], $settings['borderSizeRightTablet'], $settings['borderSizeBottomTablet'], $settings['borderSizeLeftTablet'] ), 'px' );
 		$tablet_css->add_property( 'min-height', $settings['minHeightTablet'], $settings['minHeightUnitTablet'] );
 
+		if ( ! $settings['useLegacyLayout'] ) {
+			$tablet_css->add_property( 'padding', array( $settings['paddingTopTablet'], $settings['paddingRightTablet'], $settings['paddingBottomTablet'], $settings['paddingLeftTablet'] ), $settings['paddingUnit'] );
+		}
+
 		if ( ! $settings['isGrid'] ) {
 			if ( ! $usingMinHeightFlex && $settings['minHeightTablet'] && 'inherit' !== $settings['verticalAlignmentTablet'] ) {
 				$tablet_css->add_property( 'display', 'flex' );
@@ -478,22 +489,24 @@ class GenerateBlocks_Block_Container {
 
 		$tablet_css->add_property( 'text-align', $settings['alignmentTablet'] );
 
-		$tablet_css->set_selector( '.gb-container-' . $id . ' > .gb-inside-container' );
-		$tablet_css->add_property( 'padding', array( $settings['paddingTopTablet'], $settings['paddingRightTablet'], $settings['paddingBottomTablet'], $settings['paddingLeftTablet'] ), $settings['paddingUnit'] );
+		if ( $settings['useLegacyLayout'] ) {
+			$tablet_css->set_selector( '.gb-container-' . $id . ' > .gb-inside-container' );
+			$tablet_css->add_property( 'padding', array( $settings['paddingTopTablet'], $settings['paddingRightTablet'], $settings['paddingBottomTablet'], $settings['paddingLeftTablet'] ), $settings['paddingUnit'] );
 
-		$usingMinHeightInnerWidthBoxSizing = false;
+			$usingMinHeightInnerWidthBoxSizing = false;
 
-		if ( ! $settings['isGrid'] ) {
-			// Needs 100% width if it's a flex item.
-			if ( ! $usingMinHeightInnerWidth && $settings['minHeightTablet'] && 'inherit' !== $settings['verticalAlignmentTablet'] ) {
-				$tablet_css->add_property( 'width', '100%' );
+			if ( ! $settings['isGrid'] ) {
+				// Needs 100% width if it's a flex item.
+				if ( ! $usingMinHeightInnerWidth && $settings['minHeightTablet'] && 'inherit' !== $settings['verticalAlignmentTablet'] ) {
+					$tablet_css->add_property( 'width', '100%' );
 
-				$usingMinHeightInnerWidth = true;
-			} elseif ( $usingMinHeightInnerWidth ) {
-				if ( 'contained' === $settings['innerContainer'] && ! $settings['isGrid'] ) {
-					$tablet_css->add_property( 'box-sizing', 'border-box' );
+					$usingMinHeightInnerWidth = true;
+				} elseif ( $usingMinHeightInnerWidth ) {
+					if ( 'contained' === $settings['innerContainer'] && ! $settings['isGrid'] ) {
+						$tablet_css->add_property( 'box-sizing', 'border-box' );
 
-					$usingMinHeightInnerWidthBoxSizing = true;
+						$usingMinHeightInnerWidthBoxSizing = true;
+					}
 				}
 			}
 		}
@@ -559,6 +572,10 @@ class GenerateBlocks_Block_Container {
 		$mobile_css->add_property( 'border-width', array( $settings['borderSizeTopMobile'], $settings['borderSizeRightMobile'], $settings['borderSizeBottomMobile'], $settings['borderSizeLeftMobile'] ), 'px' );
 		$mobile_css->add_property( 'min-height', $settings['minHeightMobile'], $settings['minHeightUnitMobile'] );
 
+		if ( ! $settings['useLegacyLayout'] ) {
+			$mobile_css->add_property( 'padding', array( $settings['paddingTopMobile'], $settings['paddingRightMobile'], $settings['paddingBottomMobile'], $settings['paddingLeftMobile'] ), $settings['paddingUnit'] );
+		}
+
 		if ( ! $settings['isGrid'] ) {
 			if ( ! $usingMinHeightFlex && $settings['minHeightMobile'] && 'inherit' !== $settings['verticalAlignmentMobile'] ) {
 				$mobile_css->add_property( 'display', 'flex' );
@@ -574,16 +591,18 @@ class GenerateBlocks_Block_Container {
 
 		$mobile_css->add_property( 'text-align', $settings['alignmentMobile'] );
 
-		$mobile_css->set_selector( '.gb-container-' . $id . ' > .gb-inside-container' );
-		$mobile_css->add_property( 'padding', array( $settings['paddingTopMobile'], $settings['paddingRightMobile'], $settings['paddingBottomMobile'], $settings['paddingLeftMobile'] ), $settings['paddingUnit'] );
+		if ( $settings['useLegacyLayout'] ) {
+			$mobile_css->set_selector( '.gb-container-' . $id . ' > .gb-inside-container' );
+			$mobile_css->add_property( 'padding', array( $settings['paddingTopMobile'], $settings['paddingRightMobile'], $settings['paddingBottomMobile'], $settings['paddingLeftMobile'] ), $settings['paddingUnit'] );
 
-		if ( ! $settings['isGrid'] ) {
-			// Needs 100% width if it's a flex item.
-			if ( ! $usingMinHeightInnerWidth && $settings['minHeightMobile'] && 'inherit' !== $settings['verticalAlignmentMobile'] ) {
-				$mobile_css->add_property( 'width', '100%' );
-			} elseif ( $usingMinHeightInnerWidth && ! $usingMinHeightInnerWidthBoxSizing ) {
-				if ( 'contained' === $settings['innerContainer'] && ! $settings['isGrid'] ) {
-					$mobile_css->add_property( 'box-sizing', 'border-box' );
+			if ( ! $settings['isGrid'] ) {
+				// Needs 100% width if it's a flex item.
+				if ( ! $usingMinHeightInnerWidth && $settings['minHeightMobile'] && 'inherit' !== $settings['verticalAlignmentMobile'] ) {
+					$mobile_css->add_property( 'width', '100%' );
+				} elseif ( $usingMinHeightInnerWidth && ! $usingMinHeightInnerWidthBoxSizing ) {
+					if ( 'contained' === $settings['innerContainer'] && ! $settings['isGrid'] ) {
+						$mobile_css->add_property( 'box-sizing', 'border-box' );
+					}
 				}
 			}
 		}
@@ -819,7 +838,9 @@ class GenerateBlocks_Block_Container {
 			$block
 		);
 
-		$output .= '<div class="gb-inside-container">';
+		if ( $settings['useLegacyLayout'] ) {
+			$output .= '<div class="gb-inside-container">';
+		}
 
 		$output = apply_filters(
 			'generateblocks_inside_container',
@@ -829,7 +850,10 @@ class GenerateBlocks_Block_Container {
 		);
 
 		$output .= $content;
-		$output .= '</div>';
+
+		if ( $settings['useLegacyLayout'] ) {
+			$output .= '</div>';
+		}
 
 		$output = apply_filters(
 			'generateblocks_before_container_close',
