+++ src/components/unit-control/editor.scss
@@ -3,6 +3,11 @@
 		display: flex;
 		align-items: flex-start;
 
+		.components-base-control,
+		.components-base-control__field {
+			margin-bottom: 0;
+		}
+
 		select {
 			appearance: none;
 			background: transparent;
